:sparkles: [Feat] Store Proxt add

See merge request s12-bigdata-recom-sub1/S12P21A306!194

+++ sosangomin-server/src/main/java/com/ssafy/sosangomin/api/proxy/competitor/docs/CompetitorSwagger.java
@@ -76,7 +76,7 @@ public interface CompetitorSwagger {
             }
     )
     ResponseEntity<Object> getStoreComparisonList(
-            @Parameter(description = "매장 ID") @PathVariable int storeId);
+            @Parameter(description = "매장 ID") @PathVariable Long storeId);
 
     @Operation(
             summary = "원클릭 경쟁사 분석",
